[ADDITION] Introducing the cats module
We can create the module using the command `nest g module coffees`. When we write this command, the nest cli automatically adds it to the closest module to CatsModule. In our case, the closest and the only module is AppModule.

The @Module decorator provides metadata that nest uses to structure the application.

The module decorator takes one object:

1. controller: API routes you want this module to instantiate

2. exports: Here we can list providers within this module that will be available if the module gets imported by any other module.

3. imports: imports gives us the ability to notify about other modules

4. providers: the service needs to be injected by our nest injector.