[ADDITION] Added validation to CreateCoffeeDTO and UpdateCoffeeDTO
This section is all about how we update our DTOs with Global Pipes.

As we have previously mentioned DTOs are shape that our request body should be.

However we don't have any control over the shape of DTO. As a result, we should have some sort of system built for DTO which automatically detect fault if some data is not in it's current form.

NestJS provides validation pipes to solve this exact problem. It provides a convinient way to validate all incoming client payloads.

For that we need to install two packages via this command "npm i class-validator class-transformer"