[ADDITION]
As we have seen in the previous request if we sent a POST request with the following parameters in the body "{
		"name": "Della",
		"breed": "Siamese",
		"age": 5,
		"activities": [
			"playing",
			"sleeping"
		],
	"color": "brown"
	}"

It add it to the cats, without the color, because of the whiltelabel: true.

However the expected behavior is the API to send a reverse message saying "color" is not a valid property.

We can do that just by adding "forbidNonWhiteListed: true" inside the Validation Pipe.

Now if you do the same POST request it will provide proper error message.