[ADDITION]
When we request payload, they typically come over the network as plain JS object. How can we make sure payload are coming in the shape we expect them to be.

We can do that just by implementing `transform: true`.

However, this imporvement comes with slight performance bottlenech, so check your API endpoints to be safe